back/requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==19.9.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
Pillow==8.4.0
psycopg2-binary==2.9.1
pyrsistent==0.18.0
pyTelegramBotAPI==4.1.1
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
